// src/types/location.d.ts
export type LocationsResponse = {
  info: {
    count: number
    pages: number
    next: string | null
    prev: string | null
  }
  results: Location[]
}

export type Location = {
  id: number
  name: string
  type: string
  dimension: string
  residents: string[]
  url: string
  created: string
}